+++ src/all.ts
@@ -32,6 +32,7 @@ export { CpsNotification, toast } from './components/notification.js';
 export { CpsOption } from './components/option.js';
 export { CpsProgress } from './components/progress.js';
 export { CpsRadio } from './components/radio.js';
+export { CpsRating } from './components/rating.js';
 export { CpsRadioGroup } from './components/radio-group.js';
 export { CpsSelect } from './components/select.js';
 export { CpsSeparator } from './components/separator.js';